## Changes

1. Fix still open delete-diagram modal on web version

## Issues

Closes #1579 

PR: #1580

+++ src/modules/solution-explorer/solution-explorer-solution/delete-diagram-modal/delete-diagram-modal.ts
@@ -91,15 +91,21 @@ export class DeleteDiagramModal {
       this._notificationService.showNotification(NotificationType.ERROR, message);
     }
 
-    const diagramIndex: number = this._openDiagramService
-      .getOpenedDiagrams()
-      .findIndex((diagram: IDiagram) => diagram.uri === this.diagram.uri);
+    const openDiagramServiceIsAvailable: boolean = typeof this._openDiagramService !== 'string';
+
+    const diagramIndex: number = openDiagramServiceIsAvailable
+                               ? this._openDiagramService
+                                  .getOpenedDiagrams()
+                                  .findIndex((diagram: IDiagram) => diagram.uri === this.diagram.uri)
+                               : undefined;
 
     const previousOrNextDiagramIndex: number = diagramIndex === 0 ? diagramIndex + 1 : diagramIndex - 1;
 
-    const diagramToNavigateTo: IDiagram = this._openDiagramService
-      .getOpenedDiagrams()
-      .find((diagram: IDiagram, index: number) => index === previousOrNextDiagramIndex);
+    const diagramToNavigateTo: IDiagram = openDiagramServiceIsAvailable
+                                        ? this._openDiagramService
+                                            .getOpenedDiagrams()
+                                            .find((diagram: IDiagram, index: number) => index === previousOrNextDiagramIndex)
+                                        : undefined;
 
     const diagramIsDeployed: boolean = this.diagram.uri.startsWith('http');
 
@@ -119,9 +125,11 @@ export class DeleteDiagramModal {
       });
     }
 
-    this._openDiagramService.closeDiagram(this.diagram);
-    this._solutionService.removeOpenDiagramByUri(this.diagram.uri);
-    this._openDiagramStateService.deleteDiagramState(this.diagram.uri);
+    if (openDiagramServiceIsAvailable) {
+      this._openDiagramService.closeDiagram(this.diagram);
+      this._solutionService.removeOpenDiagramByUri(this.diagram.uri);
+      this._openDiagramStateService.deleteDiagramState(this.diagram.uri);
+    }
 
     this.diagram = undefined;
     this._solutionExplorerService = undefined;
